requirements.txt
sh
pandoc
python-fasthtml
./tools/pilot/